fix: changed spelling
